hvac-backend/requirements.txt
fastapi>=0.104.0
pymodbus>=3.6.0
uvicorn>=0.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
